old-tests/submission/W3C/domhtml/HTMLObjectElement10.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/editing.html#dom-tabindex" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<object data="/support/html/sample.html" height="80" tabindex="0" type="text/html" width="550"></object>



<script>
var domtest = async_test("Index that represents the element's position in the tabbing order.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtabindex;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vtabindex = testNode.tabIndex;

      assert_equals(vtabindex,0);
       
    domtest.done();
}));

</script>
<p>Index that represents the element's position in the tabbing order.
The value of attribute tabIndex of the object element is read and checked against the expected value.</p>
<div id="log"></div>
</body></html>
